+++ desktop/cmp/dash/DashContainerModel.js
@@ -10,7 +10,7 @@ import {ContextMenu} from '@xh/hoist/kit/blueprint';
 import {GoldenLayout} from '@xh/hoist/kit/golden-layout';
 import {action, observable, bindable, makeObservable} from '@xh/hoist/mobx';
 import {wait} from '@xh/hoist/promise';
-import {PendingTaskModel} from '@xh/hoist/utils/async';
+import {AsyncTask} from '@xh/hoist/utils/async';
 import {debounced, ensureUniqueBy, throwIf} from '@xh/hoist/utils/js';
 import {createObservableRef} from '@xh/hoist/utils/react';
 import {cloneDeep, defaultsDeep, find, isFinite, reject} from 'lodash';
@@ -115,7 +115,7 @@ export class DashContainerModel extends HoistModel {
     //------------------------
     // Implementation properties
     //------------------------
-    @managed loadingStateTask = new PendingTaskModel();
+    @managed loadingStateTask = new AsyncTask();
     containerRef = createObservableRef();
     modelLookupContext;